src/VariableFinder/requirements.txt
gensim
pandas
python-levenshtein
xlrd
nltk
PyYAML
